llm-plugin-demo/backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.27.0
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
multidict==6.4.3
mypy==1.9.0
mypy-extensions==1.0.0
openai==1.75.0
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.0.1
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette
tqdm==4.67.1
typer==0.15.2
types-requests==2.31.0.20240311
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
yarl==1.19.0
